requirements.txt
PyJWT
pycryptodome
flask
pytz
icalendar
lxml
requests